style: rename session module to sessionReading module
