+++ .idea/gradle.xml
@@ -22,7 +22,7 @@
             <option value="$PROJECT_DIR$/presentation" />
             <option value="$PROJECT_DIR$/presentation/home" />
             <option value="$PROJECT_DIR$/presentation/main" />
-            <option value="$PROJECT_DIR$/presentation/session" />
+            <option value="$PROJECT_DIR$/presentation/sessionReading" />
             <option value="$PROJECT_DIR$/presentation/together" />
           </set>
         </option>
