+++ presentation/sessionReading/.gitignore
