+++ presentation/sessionReading/proguard-rules.pro
