+++ presentation/sessionReading/src/main/AndroidManifest.xml
