+++ presentation/sessionReading/src/main/res/drawable/bottom_shadow.xml
