+++ presentation/sessionReading/src/main/res/layout/activity_session_reading.xml
