+++ presentation/sessionReading/src/main/res/values/strings.xml
