+++ settings.gradle.kts
@@ -29,4 +29,4 @@ include(":core:resource")
 include(":core:navigation")
 include(":presentation:home")
 include(":core:ui")
-include(":presentation:session")
+include(":presentation:sessionReading")
